[Add] Get text from enchanted image

+++ main.py
@@ -21,7 +21,8 @@ def check_dir(*path: str):
 pytesseract.pytesseract.tesseract_cmd = r"C:\Program Files\Tesseract-OCR\tesseract.exe"
 
 # Clear all files and recreate
-shutil.rmtree("working_folders")
+if os.path.exists("working_folders"):
+    shutil.rmtree("working_folders")
 check_dir(
     "working_folders/output_slices", 
     "working_folders/bubbles", 
@@ -72,14 +73,43 @@ des_path = "working_folders/enchanced_images"
 
 for sub_dir in os.listdir(path):
     for in_sub in os.listdir(os.path.join(path, sub_dir)):
-        print(in_sub)
+        os.makedirs(os.path.join(des_path, sub_dir, in_sub), exist_ok=True)
+        subprocess.run([
+            "Real-ESRGAN/realesrgan-ncnn-vulkan.exe",
+            "-i", os.path.join(path, sub_dir, in_sub),
+            "-o", os.path.join(des_path, sub_dir, in_sub),
+            "-n", "realesrgan-x4plus",
+        ])
 
-# for sub_dir in os.listdir(path):
-#     os.makedirs(os.path.join(des_path, sub_dir), exist_ok=True)
-#     subprocess.run([
-#         "Real-ESRGAN/realesrgan-ncnn-vulkan.exe",
-#         "-i", os.path.join(path, sub_dir),
-#         "-o", os.path.join(des_path, sub_dir),
-#         "-n", "realesrgan-x4plus"
-#     ])
 
+# Get text from enchanced image
+count = 0
+succ_count = 0
+
+langs = "kor+eng+ind"
+pytesseract.pytesseract.tesseract_cmd = r'C:\Program Files\Tesseract-OCR\tesseract.exe'
+
+for sub_dir in os.listdir(path):
+    print(f"\nsub_dir: {sub_dir}")
+
+    for in_sub in os.listdir(os.path.join(path, sub_dir)):
+        print(f"\nin_sub: {in_sub}")
+        in_sub = os.path.join(path, sub_dir, in_sub)
+
+        for img in os.listdir(in_sub):
+            image = cv2.imread(os.path.join(in_sub, img))
+            # gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+
+            text = pytesseract.image_to_string(image, lang=langs, config="r'--oem 3 --psm 7'")
+            text = text.replace("\n", '')
+
+            print(f"{img}: ", end='')
+            print(text)
+
+            if len(text) != 0:
+                succ_count += 1
+            count += 1
+
+    print("-" * 50)
+
+print(f"Result: {succ_count}/{count} can read {succ_count * 100 / count :.2f}% in this comic")
